@@ -5,8 +5,8 @@
   "license": "MIT",
   "dependencies": {
     "final-form": "^4.0.0",
-    "vue": "^2.0.0",
-    "vue-router": "^3.0.0",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0",
     "vue-final-form": "latest"
   }
 }
