@@ -14,27 +14,24 @@
     "test": "npm run lint && echo 'no tests!'",
     "lint": "xo",
     "prepublishOnly": "npm run build",
-    "build": "bili --format umd,umd-min --module-name VueFinalForm --global.final-form final-form --sizeLimit.umd-min 1.5kb",
-    "example": "poi --serve",
-    "build:example": "poi --prod",
+    "build": "bili --format umd,umd-min --module-name VueFinalForm --global.final-form final-form --global.vue vue",
+    "example": "vite",
+    "build:example": "vite build",
     "gh": "gh-pages -d example/dist",
     "deploy": "npm run build:example && npm run gh"
   },
   "author": "egoist <0x142857@gmail.com>",
   "license": "MIT",
-  "dependencies": {
-    "nano-assign": "^1.0.0"
-  },
   "devDependencies": {
-    "bili": "^3.4.2",
-    "eslint-config-rem": "^3.0.0",
+    "@vitejs/plugin-vue": "^1.10.0",
+    "bili": "^5.0.5",
+    "eslint-config-rem": "^4.0.0",
     "final-form": "^4.0.0",
     "gh-pages": "^1.0.0",
-    "poi": "^12.0.0",
-    "vue": "^2.5.21",
-    "vue-router": "^3.0.1",
-    "vue-template-compiler": "^2.5.21",
-    "xo": "^0.18.0"
+    "vite": "^2.0.0",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0",
+    "xo": "^0.47.0"
   },
   "xo": {
     "extends": "rem",
